+++ FizzimGui.java
@@ -2339,6 +2339,13 @@ public class FizzimGui extends javax.swing.JFrame {
 		propertyInspectorEditButton.setEnabled(true);
 		if(obj.getType() == 3)
 		{
+			if(((TextObj)obj).getGlobalTable())
+			{
+				propertyInspectorTable.setModel(new ReadOnlyInspectorTableModel(
+						new Object[][] {{"Text", "FSM interface summary"}},
+						new Object[] {"Field", "Value"}));
+				return;
+			}
 			propertyInspectorTable.setModel(new ReadOnlyInspectorTableModel(
 					new Object[][] {{"Text", ((TextObj)obj).getText()}},
 					new Object[] {"Field", "Value"}));
@@ -2417,6 +2424,11 @@ public class FizzimGui extends javax.swing.JFrame {
 		}
 		else if(obj.getType() == 3)
 		{
+			if(((TextObj)obj).getGlobalTable())
+			{
+				openGlobalPropertiesTab(0);
+				return;
+			}
 			drawArea1.editText((TextObj)obj);
 		}
 		updatePropertyInspector(drawArea1.getSelectedObjectsForInspector());
@@ -2942,6 +2954,14 @@ public class FizzimGui extends javax.swing.JFrame {
 		new GlobalProperties(drawArea1, this, true, globalList, 0)
 				.setVisible(true);
 	}//GEN-LAST:event_GlobalItemMachineActionPerformed
+
+	public void openGlobalPropertiesTab(int tab) {
+		if(tab < 0)
+			tab = 0;
+		globalList = drawArea1.setUndoPoint();
+		new GlobalProperties(drawArea1, this, true, globalList, tab)
+				.setVisible(true);
+	}
 	
 	private void GlobalItemInputsActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_GlobalItemsInputsActionPerformed
 		globalList = drawArea1.setUndoPoint();